streamlit/requirements.txt
albumentations==1.3.0
altair==5.0.1
anyio==3.7.0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.2
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==4.0.1
beautifulsoup4 @ file:///home/linux1/recipes/ci/beautifulsoup4_1610988766420/work
bleach==6.0.0
blinker==1.4
boto3==1.26.144
botocore==1.29.144
brotlipy==0.7.0
cachetools==5.3.1
certifi==2021.10.8
cffi @ file:///tmp/build/80754af9/cffi_1605538068321/work
chardet @ file:///tmp/build/80754af9/chardet_1605303185383/work
charset-normalizer==3.1.0
click==8.1.4
cmake==3.26.3
comm==0.1.3
conda==4.9.2
conda-build==3.21.4
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1603018141399/work
contourpy==1.0.7
cryptography==41.0.0
cycler==0.11.0
debugpy==1.6.7
decorator==4.4.2
defusedxml==0.7.1
dnspython==2.1.0
efficientnet-pytorch==0.7.1
exceptiongroup==1.1.1
executing==0.8.3
fastjsonschema==2.15.3
filelock @ file:///home/linux1/recipes/ci/filelock_1610993975404/work
fonttools==4.39.4
fqdn==1.5.1
fsspec==2023.6.0
gevent==22.10.2
gitdb==4.0.10
GitPython==3.1.32
glob2 @ file:///home/linux1/recipes/ci/glob2_1610991677669/work
greenlet==2.0.2
h5py==3.9.0
huggingface-hub==0.16.4
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
imageio==2.30.0
importlib-metadata==6.6.0
importlib-resources==5.12.0
inotify-simple==1.2.1
ipykernel==6.23.1
ipython==8.12.2
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets==8.0.6
isoduration==20.11.0
jedi @ file:///tmp/build/80754af9/jedi_1606932564285/work
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.14
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
jupyterlab_server==2.22.1
kaggle==1.5.15
kiwisolver==1.4.4
lazy_loader==0.2
libarchive-c @ file:///home/linux1/recipes/ci/python-libarchive-c_1610974153025/work
lit==16.0.5
markdown-it-py==3.0.0
MarkupSafe==2.0.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==0.8.4
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
mpmath==1.3.0
munch==4.0.0
nbclient==0.8.0
nbconvert==7.4.0
nbformat==5.9.0
nest-asyncio==1.5.6
networkx==3.1
notebook_shim==0.2.3
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
olefile==0.46
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
opendatasets==0.1.22
overrides==7.3.1
packaging==23.1
pandas==2.0.2
pandocfilters==1.5.0
paramiko==3.2.0
parso==0.7.0
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.5.0
pkginfo==1.7.0
pkgutil_resolve_name==1.3.10
platformdirs==3.5.1
pretrainedmodels==0.7.4
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil @ file:///tmp/build/80754af9/psutil_1598370257551/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval==0.2.2
pyarrow==12.0.1
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.2
python-etcd==0.4.5
python-json-logger==2.0.7
python-slugify==8.0.1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==5.3.1
pyzmq==22.3.0
qudida==0.0.4
requests==2.31.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
ruamel_yaml==0.15.87
s3transfer==0.6.1
safetensors==0.3.1
sagemaker-training==4.5.0
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
seaborn==0.12.2
segmentation-models-pytorch==0.3.3
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
streamlit==1.24.1
sympy==1.12
tenacity==8.2.2
terminado==0.13.3
text-unidecode==1.3
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.9.2
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
torch==1.11.0
torchelastic==0.2.1
torchvision==0.12.0
tornado==6.1
tqdm==4.65.0
traitlets==5.2.0
triton==2.0.0
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.3.1
ultralytics==8.0.135
uri-template==1.2.0
urllib3==1.26.9
validators==0.20.0
watchdog==3.0.0
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.2
Werkzeug==2.3.4
widgetsnbextension==4.0.7
zipp==3.8.0
zope.event==4.6
zope.interface==6.0
